Refactor and consolidate Selenium test scripts; remove outdated files and enhance test coverage for user features.
